Create MovieRating.sql
Write a solution to:
Find the name of the user who has rated the greatest number of movies. 
In case of a tie, return the lexicographically smaller user name.

Find the movie name with the highest average rating in February 2020. 
In case of a tie, return the lexicographically smaller movie name.